+++ packages/debug/src/node/stream-debug-adapter.ts
@@ -19,7 +19,7 @@ import { Emitter, Event } from '@theia/core/lib/common/event';
 import { ChildProcess } from 'child_process';
 import * as stream from 'stream';
 import * as net from 'net';
-import { DebugAdapter } from './debug-model';
+import { DebugAdapter } from '../common/debug-model';
 
 abstract class StreamDebugAdapter extends DisposableCollection {
     private messageReceivedEmitter = new Emitter<string>();